Use oslo_db fixtures
Converts the plain and migration fixtures to use oslo_db
test_fixtures.

The one snag that can be improved on the oslo_db side is
that the generate_schema hooks called by test_fixtures.SimpleDBFixture
are called before the enginefacades have been connected with the
provisioned engine, so an additional patch is called here
as placement's database setup relies upon enginefacade-enabled
methods to insert DB data.

The WalkVersionsMixin here is necessary as the one in
oslo.db is specific to SQLAlchemy-migrate, so oslo.db should
also include this alembic-capable WalkVersionsMixin within
its own modules.

The normal provisioning system of generating anonymously-named
databases within MySQL and Postgresql is in place, which allows
tests to run concurrently in multiple processes without conflicts.

Change-Id: Icf2ea4a5f51dce412c6e93fda3950d0c9f17aa9f