Add stamp DB version to the migration script
The existing nova-api to placement database migration shell scripts
don't include the post task to stamp the placement database to the
(b4ed3a175331) version. Without that stamping, syncing the DB to the
current version fails on having the initial upgrade: (initial) ->
(b4ed3a175331). This is because it assumes the initial DB has no
contents and tries to create the tables again.

This patch changes the shell scripts to include that stamping task
rather than leaving it as operators' manual duties to safely bring
the placement DB under alembic version control.

With this change, the shell scripts will need to be executed under the
following condition:
 - The placement is already installed to execute `placement-manage *`
 - The placement cli can access the placement's database, for example,
   by reffering to the `[placement_database]` section in the
   `placement.conf`.

Depends-On: https://review.openstack.org/620485
Change-Id: I75926b0efb3983d62603f2fd30b5a8cc30203d46