Add irrelevant files list to perfload job
We don't want perfload to run on docs, tox, or test changes, but
because it is based off a very basic job (to avoid weight) there was
no built in irrlevant-files list. This adds one to ensure that the
job only runs on code-related changes.

We need a slightly different list of files from the list used by
integrated gate jobs so do not use the changes made in
Iae26251cb455640fbc74dad326666d2015e7a46c .

Change-Id: I092b5ed19ef51c42a714984127e2fdc91d5c6c70