Retry _ensure_aggregates a limited number of times
Use the oslo_db provided wrap_db_retry to retry _ensure_aggregates a
maximum of 5 times when DBDuplicateEntry happens.

This replaces previous looping which had no guard to prevent unlimited
looping if for some reason DBDuplicateEntry kept happening. This is
possible, rarely, when there are is a small number of threads talking
to the same database, and a small number of aggregates of aggregates
being ensured a huge number of times (thousands) in a very small amount
of time. Under those circumstances a maximum recursion error was
possible.

The test of the new behavior is shamefully over-mocked but manages to
confirm the behavior against the real database, thus is in the
functional tree.

Change-Id: I67ab2b9a44264c9fd3a8e69a6fa17466473326f1
Closes-Bug: #1804453