Remove dead code in objects/resource_provider.py
Coverage testing revealed that a few methods in
objects/resource_provider.py had no coverage. It was initially assumed
this was an artifact of the extraction of placement from nova.

Turns out that the code is not used. It was added in
Ib45fde56706a861df0fc048bbec8a568fd26d85d but then refactoring in
Ib1738fb4a4664aa7b78398655fd23159a54f5f69 made it redundant but the
changes did not remove the unused code.

Change-Id: Id2f0a85b58f191886a2edea8fc195bc2289cbcec
Closes-Bug: #1805858