Document API error codes
Add a section to the api-ref describing the error codes that some
responses produce.

Note in the contributor docs that this should be updated when one is
added.

The reshaper docs is adjusted so a ref can be made to it from the
errors. The implicit link to the header that would be the norm there
doesn't work as there are two headers named "Reshaper".

Change-Id: I89bbd383ba102fdd707ccc9f2fc973c6dd841fa8
Closes-Bug: #1794712