Also time placeload when doing perfload
The previous iteration was only timing how long it took to GET some
resource providers after we create 1000 of them.

It's also useful to know how long it takes to create them.

Neither of these timings are robust because we do not have reliable
sameness from virtual machine to virtual machine (especially between
cloud providers) but they make it possible to become aware of
unusual circumstances.

To avoid extraneous noise in the placement-perf.txt file, set +x
and set -x surrounds the commands that create that output.

Change-Id: I4da2703dc4e8b306d004ac092d436d85669caf0f