Placement install documentation
This is a first pass at creating the install documentation for
placement, by copying the nova install documentation, and removing the
nova parts, leaving the placement parts in place.

The expectation here is that this provides a starting point which
can be iteratively improved, perfection being the enemy of the
done and all that.

The from-pypi and verify pages are currently left as TODO and not
included in the visible table of contents. They are present in
the hidden table of contents.

Change-Id: I2f7bcd8efabc628bd27e3a9ce74e277a9e37fb69