Trim the release notes to just stein/master
This removes all the releasenotes related to changes already released
with nova in rocky and prior. They were carried over into placement
as we weren't sure how we were going to manage release notes. In
a meeting it was decided that we would start the release notes fresh
with "since-extraction".

Therefore this change removes all the releasenotes that are already
present in nova, keeping anything new.

The index.rst file is already set to say "look at nova" for older
release notes.

Change-Id: If9255212acf21ed69abbed0601c783ad01133f72