Copy create_incomplete_consumers online data migration from nova
Change Id609789ef6b4a4c745550cde80dd49cabe03869a in nova added
online data migration code to create missing consumer records
for old allocations. That was added to nova in Rocky. The
incomplete consumers are already migrated in the REST API when
showing allocations for a given consumer or listing allocations
for a given resource provider.

This adds the online data migration command line entry point
for "placement-manage db online_data_migrations" so that
incomplete consumer records can be migrated on-demand in batches
by the operator.

The nova change had no diret testing of the CLI entry point since
it's just a call to the same code that the API uses, which is already
tested in CreateIncompleteConsumersTestCase, so no explicit CLI
unit test is added here since it would be redundant.

This is part of the placement extraction effort; in the case that
a deployment migrates to extracted placement before completing
the online data migration in nova, this allows them to still complete
the migration within extracted placement.

Change-Id: If5babb29b13a3e8c26ac04ecee02f4d3d5404263