Increase loop size on _ensure_aggregate
The loop is used to cover when more than one different thread of
activity is trying to add the same aggregate. Initially set at
five, perfload results [1] are showing that this is not high
enough. While it is true that perfload exercises writing to
placement in ways that are unusually concurrent, we want to
be robust and enlarging the loop size doesn't impact the common
case, so it has been raised to ten.

[1] http://logs.openstack.org/99/632599/1/check/placement-perfload/8c2a0ad/logs/

Change-Id: I288fc07df30adb044e0ce1614519b3391cf9bb54