Adjust database connection pool config in perfload tests
The perfload tests can run out of connections in the sqlalchemy
connection pool when using the default configuration. This can
lead to distracting noise in the results [1] and potentially
failures. Since it is easy to adjust the settings for the job,
let's do that.

The perfload web service is set up for enabling quite wide
concurrency, so the database connections need to be as well.

[1] http://logs.openstack.org/99/632599/1/check/placement-perfload/8c2a0ad/logs/

Change-Id: Id88fb2eaefaeb95208de524a827a469be749b3db