Use one ConfigOpts in placement-status
Because of a conflict between how the oslo.upgradecheck library uses
oslo config and how we want to use it in placement, two different
ConfigOpts were needed to avoid an args already parsed error. The
new release of the oslo.upgradecheck library is change to allow
the two steps in main (registering the CLI opts and running the
command) to be called separately if desired. Doing so allows us
to use just one ConfigOpts.

Requirements and constraints are updated.

Change-Id: I792df18cb17da95659628bfe7f7a69897c6f37ab