Don't use OVO with allocation candidates
I did some fairly basic benchmarking a while back which indicated
that apart from time spent in the database, the value coercing
and type checking that happens in OVO and the associated getters
and setters accounts for a fair bit of time and a rather large
number of the function calls.

This change and its children explores if using (very) basic objects
instead of OVO will work and be of any benefit. In this patch,
allocation candidates and the objects nested within are modified.

Benefit can be defined in a variety of ways:

* Is more performant
* Is more maintainable
* Doesn't remove functionality

Things to watch out for/question/be aware of:

* None of the objects involved here need a context member. Wherever
  context is used it is supplied by the callers.
* Type checking is gone. In this particular context it shouldn't
  normally matter: there's no external interface to these objects.
  An argument could be made that the type checking is of use to
  developers, but we have type checking at both the DB and HTTP
  API levels (and tests, of course). At the intermediary
  provided by the objects it might be noise, especially if the
  performance impact is noticeable.
* Type coercing is sometimes important. For example values that
  are the result of sql query including a func.sum will be
  presented as a Python Decimal. If this value makes it to the
  json serializer, json will barf. A change is included here
  for the 'used' value that shows up in allocation candidate
  results. Note that this problem doesn't show up in functional
  tests, just tempest and grenade, because it needs MySQL to
  happen.
* If we wanted to do, this could go further, to named tuples, attrs,
  or a superclass which does the field management, but so far that seems
  overkill for the classes changed thus far. Sticking simple classes
  gives us the explicit functionality we want without much in
  the way of overhead.

Patches following this one change other OVO-based objects. Once they
are all changed, patterns of commonality will be analysed to drive
some refactorings and cleanups to limit duplication.

Change-Id: I765fef25120204ac364e6f9b0343f1bda8ac86fe