Don't use OVO in Trait and TraitList objects
Turn Trait and TraiList into classical python objects.

The changes here are nearly identical to those made to
ResourceClass and ResourceClassList which indicates some
opportunities for less duplicated code.

Some of the duplication is okay, and helps to keep things
explicit, but some of it is extraneous.

Interesting note: there was still a remotable method in here.

Change-Id: I954c6d0a86fc45dc689a2c54a3d34382dfd68bf5