Factor listiness into an ObjectList base class
While de-OVO-ifying objects, all the *List classes ended up with a
common pattern of __init__, __len__, and __getitem__ methods to make
them behave listily. Here we factor these out into a base ObjectList
class in a new placement.objects.common module.

Note that there's still commonality among the *List classes that could
be factored into ObjectList - e.g. the _set_objects method. That is left
for a future change.

AllocationList was previously saving a context arg as an instance
variable and using it in replace_all and delete_all. All calling paths
of those methods had a context in scope that was the same as the one
used to init the AllocationList object itself. This change adds a
context arg to replace_all and delete_all, allowing us to remove the
instance variable. This allows us to remove the __init__ method from
AllocationList, since it can now simply use the one from the ObjectList
superclass.

Change-Id: Ice41cbdcec62d0e7d0f058bcae677e416189b1ae