Move *List.__repr__ into ObjectList
Continue to DRY listy classes by moving the __repr__ method into the
ObjectList base class.

Note: unlike previous refactors, this is not a no-op change, as it will
affect the repr() of some *List classes that didn't previously override
__repr__. Since the repr is just used to make pretty logs/errors, this
should be copacetic. (Besides, I'm pretty sure we wildly changed the
default __repr__ when we de-OVO-ified.)

Also note: For lists of objects that are especially long this could
result in very long log messages, which can cause issues with some
logging tools. We will want to watch this and explore other options.

Change-Id: Ia35403c5fe2afefd9b8d077b36561963abe16b34