Clean up ObjectList._set_objects signature
Defining ITEM_CLS on *List classes allows us to be explicit (more than
just the naming convention) about what kind of thing is being List'd. It
also lets us do away with the argument that was that class.

_set_objects was also taking an instance of the *List class itself,
which we can do away with by making _set_objects a @classmethod
instead of a @staticmethod.

Change-Id: I76039ef9f3c62713577c58b45399efdd79953762