Downgrade os-traits/os-resource-classes sync log level to DEBUG
The current count of traits from the os-traits library is 120 so
the INFO log message to dump the traits that get synchronized into
the traits table on startup of the service is quite large.

This drops that logging from INFO to DEBUG and does the same for
os-resource-classes which is admittedly smaller than os-traits but
still debug information more than informational.

Change-Id: Ib78c326d79b1ad2393c42702f9081a7593097e3e
Related-Bug: #1813147