Set timestamps in Allocation objects
The following methods return an AllocationList object
which contains a list of Allocation objects that
do not have the 'created_at' and 'updated_at' values.

* get_all_by_resource_provider in AllocationList object
* get_all_by_consumer_id in AllocationList object

It causes wrong last-modified response headers
in the following APIs.

* GET /allocations/{consumer_uuid}
* GET /resource_providers/{uuid}/allocations

So set the 'created_at' and 'updated_at' values of
the Allocation objects in the methods.

Change-Id: I9b3a990d3c635cf09e2016bda0da9cc4fb395873
Closes-Bug: #1816230