Cast Usage.usage to int
When using mysql (and perhaps postgresql, but not sqlite), the
func.sum functionality provided in sqlalchemy will return a result
that is a decimal.Decimal. The json module is unable to cope
with this. It needs "native" types of int or float.

In the case of Usage we want an int. Since there are two places
where the data is loaded, but both then directly call to create
a Usage class, do the casting in the constructor of Usage. A
unit test of that cast is added.

Since the problem does not show up in sqlite, no functional test
is added, but patches which follow this one which add integration
tests using gabbi will fail without this change in place, so if
they are passing that services as additional verification that
this is a good change.

Change-Id: I6de32d0d06b6e2427ae0cf52ed24fd0173b74baf
Closes-Bug: #1817633