Docs: extract testing info to own sub-page
In preparation for extending the testing documentation to include
information about more than just gabbi, including the tempest,
grenade and perload jobs, pull the content out to its own page.

As noted in a TODO within, making subpages show up is cumbersome
for the moment, but we can iterate to something swell.

Change-Id: I71d000804985aaf5f3a1c467974ac7f8d0250a98