Retry new transaction on failure
At high concurrency of setting aggregates, _ensure_aggregate() could
raise a DBDuplicateEntry error. The oslo_db_api wrapper caught it and
retried the same function, but it couldn't find the new entry again
because it tried to find it in the same transaction.

This blindness of the new record could happen depending on the DB
transaction isolation level, but it happens with "REPEATABLE_READ"
level, to which MySQL defaults.

This patch moves the retrying wrapper to the upper layer,
_set_aggregate(), to restart a new transaction when the
DBDuplicateEntry error happens so that it can find the new entry
another transaction has added.

Change-Id: I85ebc98502e9ca694ff07746e8e23c26e386df3c
Closes-Bug: #1818498