Move Allocation and AllocationList to own module
This is a precursor to getting rid of the AllocationList
class in favor of operating on native python lists.

When objects/allocation.py exists and is well-formed, the next
step will be to get rid of AllocationList similar to [1] but more
complex as there are write operations to deal with.

[1] I0066fc8fa9b61196ea99e35e780a7d123612023f

Change-Id: Ic673dfa95f6c8e42692c6287fb8d7cc27ee6cbf6