Use native list for lists of Allocation
Remove the AllocationList class and make the retrieval and
replace methods on it module level in placement.objects.allocation.

The methods that perform write operations on a list of allocations
take the list as an argument.

Adapting to this change requires a lot of changes in tests and
the allocation handler.

Change-Id: I2b59156e6f376df630bb0b574682964428c65c04