Add explicit short-circuit in get_all_by_consumer_id
PyCharm complained of a potentially-unset local variable in
get_all_by_consumer_id, even though that wasn't actually possible given
the list comp would be iterating over zero values in that case. However,
adding an explicit short-circuit for this case makes the code clearer to
read, and gets rid of the IDE warning, so \o/.

An unrelated typo in a comment is fixed at the same time.

Change-Id: Ic7b5a223a8eab8f7dd651521b8723d89abf2ded1