Make the PlacementFixture usable without intercept
There are cases where it is necessary to use the PlacementFixture
without it running a wsgi-intercept, for example when you'd like
to run an indepdendent wsgi-intercept, as in the nova tests that
exercise the report client.

In the bug listed as partial below those report client tests were
using a nova conf global to initialize the wsgi intercept. They
just happened to be working because that CONF has the required
configuration settings. When placement's conf diverged [1] things
broke. A nova change which depends on this one will resolve the
bug.

Note that as a part of this change, the code which deletes
'sync_on_startup' (the new conf option) before configuring the
database, is adapted to be more flexible. There really is only
one conf setting that is required to start the placement database,
connection, and we'd like to make it possible for people to set
just that when doing tests or experiments and not have to worry
about extraneous optional things.

Some parameter descriptions are added to the PlacementFixture
__init__.

We should add more robust testing to all three of PlacementFixture,
fixtures.Database, and PlacementDirect, but at the moment we are
trying to fix a gate blocking bug.

Partial-Bug: #1818560

[1] Ie43a69be8b75250d9deca6a911eda7b722ef8648

Change-Id: I8c36f35dbe85b0c0db1a5b6b5389b160b68ca488