Move ResourceClass and ResourceClassList to own module
This is a precursor to getting rid of the ResourceClassList
class in favor of operating on native python lists.

With the move, also create new unit and functional test files
for the new module, and move resource class syncing into the new
module.

The ensure_resource_classes_sync is renamed to ensure_sync since
its membership in the resource_class module gives sufficient
namespacing.

Change-Id: I95c383e22f11fce2eb68ef29106fe045315644b6