Move Trait and TraitList to own module
This is a precursor to getting rid of the TraitList
class in favor of operating on native python lists.

With the move, also create new unit and functional test files
for the new module.

The contents of trait.py is a copy of what had been in
resource_provider.py with some changes that may be of note:

* classes are put before methods, public methods before private
* a direct import of sqlalchemy's null() is replaced by using
  the already imported sa.null(). It's the same thing.
* ensure_trait_sync renamed to ensure_sync.
* get_trait_ids_from_names becomes ids_from_names
* ResourceProviderTraitTestCase is renamed to TraitTestCase

Change-Id: Ibe11ebd7ccc97a46586edd6c897b173ed67ca134