Update get trait and traits calls that were writers to reader
There were initially writer because there had been an
ensure_trait_sync (no ensure_sync) in them some time ago.
When trait sync was moved to the start of the placement
web service process [1], the writers were not corrected to
a reader.

A related comment on _trait_sync is updated to reflect the
current syncing reality.

[1] Ic87518948ed5bf4ab79f9819cd94714e350ce265

Change-Id: I8f9413b346f6da221b4fff2d37980f00d82e3e3a