Move allocation candidate classes and methods
The AllocationCandidates class and the classes it uses to assemble
a suite of classes are moved to objects/allocation_candidate.
Also included are the several methods used solely by those classes.

Methods which are more directly about resource providers, even
if only used by the allocation_candidate modules, are kept in
resource_provider.py. As a result several methods in
resource_provider.py are made public.

Tests are updated accordingly.

As this is already a rather chaotic change, no reordering was
done of the classes and methods in allocation_candidate. That
will follow.

Change-Id: I6539d7db87bc24b03842e3d28a4296002b3549d5