Move Inventory and InventoryList to own file
The Inventory and InventoryList classes are moved to an
inventory.py file. Tests and other callers are udpated
to reflect the new location.

A new unit.objects.test_inventory is created. A
functional.db.test_inventory is _not_ created because though
there are inventory related tests within test_resource_provider
they are within a ResourceProvider-related class and teasing
out the ones that are for the sake of just Inventory versus
for both Inventory and ResourceProvider is not entirely
clear.

Change-Id: Ia1ad18585194236647ec2c83480254531a619e91