Do not separately configure logging for alembic
Before this patch, when calling alembic commands (as done for db sync,
online migrations and the automatic sync when
[placement_database]/sync_on_startup is True) the env.py file would be
evaluated, clobbering the logging configuration already set in the
process. In the command line tools we didn't notice this because
there's not much logging, but when it happens in the web-service process
it means that while the process may start out logging at a DEBUG level
(if you asked for that) it would then stop once the db sync happened.

It turns out that we can remove the call to fileConfig in env.py and
get rid of the associated configuration settings in alembic.ini
we still get the behaviors we want from the placement-manage
and alembic command line tools.

Change-Id: Id9bc515cee71d629b605da015de39d1c9b0f8fc4
Story: 2005187
Task: 29932