Remove ResourceProviderList class
The get_all_by_filters class method is now module level. Callers
and tests are updated accordingly.

As there is no longer any package using common.py, it is removed.

Looking at this, it now seems obvious that 'get_all_by_filters' is
oxymoronic but changing it now didn't seem appropriate. We may
wish to do that in future work.

Change-Id: Ib9365200ebec58002c2456435f858466c6afea7f