Update CONTRIBUTING and HACKING
Update README.rst to not point to HACKING.rst, only CONTRIBUTING.rst.

Remove HACKING.rst because it is redundant in purpose, and incorrect
in content.

Update CONTRIBUTING.rst to point to the contributor docs on
docs.openstack.org and generally be a bit less generic than before.

Change-Id: Ib608b494e027091ae324793f275bad3b554e9710
Story: 2005190
Task: 29949
Task: 29950