Remove unused table constants from allocation_candiate.py
The constants are not used in allocaion_candidate.py and
are not referenced from elsewhere. The reason they ended
up in there was because during the refactoring at [1] there
was indecision about whether some methods should be in
resource_provider.py or allocation_candidate.py. They
ended up in resource_provider.py but the list of tables
was not updated.

[1] I6539d7db87bc24b03842e3d28a4296002b3549d5

Change-Id: I3d501543f7e79a8c3c030369af1398ca68f2b811