Extract user and endpoint creation install doc
The placement user and service catalog endpoint creation information
in the three different distro docs is very nearly exactly the same
so it has been extracted to a file which is included by all three.
This file can then be reused elsewhere as well, such as planned
"from-pypi" install docs (for which there exists a stub).

The only difference in the three versions was the URL used for
the endpoint(s) which already required some cooking to work in
any environment. A note has been added to point out the installar's
responsibilities in this case.

Subsubheadings are introduced in the surrounding document so that
the numbered list is considered two separate lists. Otherwise the
output ends up looking like a single list with two initial number
one items.

Change-Id: Ib928162702a54f000161367c90172c730cb632c7