Slightly improve usage documentation
There wasn't really a good overview on how to use placement,
so I've tried to start something by expanding on how nova uses
placement in very broad strokes. The intent here is not to
instruct people on the details of how to manage resources with
placement, rather that they _can_ manage resources with
placement and here's the major steps involved.

Change-Id: I11cd622162efce2e87a21fa5ae5d297f4b248541