Rename and restructure install docs
The concept/name of controller was copied over from nova when the
docs were imported. That concept doesn't really fit in with
placement and also does not describe what the docs are actually
doing. We're not installing a controller, we're installing
placement.

Thus filenames and titles are updated and intro paragraphs
clarified.

Also, the endpoints.rst file is moved into a shared directory as it
is not a standalone install doc.

Note that the big warning remains on the top of the install docs
because it is still true. These instructions and the packages they
mention have not been verified.

Story: 2005190
Change-Id: I5fcbc90eda3ef74dba2336ea3e5c9f53938c6378