Update the from-pypi install instructions
Provide step by step instructions for installing from pypi.
Note that this is a greenfield installation, not an upgrade.

Only one of many possible web servers is described as there
are too many choices to describe them all.

The depends-on adds oslo.log to config doc generation,
allowing a link to the debug option. This is done as a
depends-on to avoid a rebase.

Depends-On: https://review.openstack.org/644589
Change-Id: I580fa4394cb93b8e8141ee2d546543c174356a47
Story: 2005190
Task: 29943