Initial structure for in-tree specs
We decided in IRC and email discussions that henceforth
placement specs will reside in the same tree as code and docs.
This change sets that up using the following structure:

* There is a docs/source/specs top-level index that should be
  updated with each release to include tables of contents generated
  from cycle directories (currently pointing to train) for
  approved and implemented specs.

* In the approved directory for any cycle there will be a
  <cycle-template>.rst file to use as the basis for any new
  spec. New specs go in this directory.

* In the implemented directory is a placeholder file which
  will be removed when the first (if any) spec moves from
  approved (meaning in-progress) to implemented. The PTL
  or their designee will be responsible for doing that moving.
  The placeholder file is required to avoid warnings (which
  will kill the tox docs job) from the table of contents
  globbing.

* The train-template.rst is a copy-with-edits from nova's
  version of the same thing. The main changes are to indicate
  that StoryBoard, not launchpad blueprints, are the starting
  point and to remove various nova-isms.

  It is likely we will want to further tweak the template as
  we decide what matters.

The next steps for this are for people with placement specs
that had been pending in nova in stein, to resubmit them to
placement as train specs (in doc/source/specs/train/approved).
It is quite likely and not surprising that it will sometimes
be confusing whether a spec should be in nova or placement.
Ideally we can identify the "just placement" aspects of a
change and make a placement-only spec for that which the nova
spec makes reference to. This may sound like a pain but it is
a feature, not a bug, which helps make sure that placement
evolves its API as a strong contract for all potential clients.

Change-Id: I85854d771409701cea2fcf7a218f02af60dba72e